requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cssselect==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.9.6
lxml==4.1.0
MarkupSafe==1.0
newspaper3k==0.2.5
nltk==3.2.5
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
requests-file==1.4.2
six==1.11.0
tldextract==2.1.0
urllib3==1.22
Werkzeug==0.12.2
pytest==3.2.3
